requirements.txt
pillow
numpy
flask
gunicorn
requests